PR feedback
- Single class per file
- Don't use swap for move, because there is no particularly compelling
  reason for it
- provide namespace swap instead of template specialization
